Move ErrorBoundary component out of 'atoms' folder
