+++ packages/client/src/components/ErrorBoundary.tsx
