+++ packages/client/src/pages/error_boundary/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { LayoutContent } from "@eisbuk/ui";
 
 import Layout from "@/controllers/Layout";
-import ErrorBoundary from "@/components/atoms/ErrorBoundary";
+import ErrorBoundary from "@/components/ErrorBoundary";
 
 const ComponentThatThrows = () => {
   throw new Error("This is an error");
